--- lib/bowler.rb
@@ -1,11 +1,0 @@
-class Bowler
-  def set_name(string)
-    puts "Setting name.."
-    @name = string
-  end
-
-  def get_name
-    puts "Returning name.."
-    @name
-  end
-end
