+++ lib/game.rb
@@ -1,7 +1,7 @@
 class Game
 
   PINS = 10
-  FRAMES = 10 
+  FRAMES = 10
 
   def initialize
     @rolls = []
@@ -12,38 +12,29 @@ class Game
   end
 
   def score
-    result = 0
-    frame_index = 0
+    result = 0 
+    frame_index = 0 
 
     FRAMES.times do
-      if strike?(frame_index)
-        result += strike_bonus(frame_index)
+      if @rolls.fetch(frame_index,0) == PINS #strike
+        result += PINS + @rolls.fetch(frame_index + 1, 0) + @rolls.fetch(frame_index + 2, 0)
         frame_index += 1
       elsif spare?(frame_index)
-       result += spare_bonus(frame_index)
-       frame_index += 2
-     else
-       result +=  @rolls.fetch(frame_index,0) + @rolls.fetch(frame_index + 1,0)
-       frame_index += 2
-    end
-      
+        result += spare_bonus(frame_index)
+        frame_index += 2
+      else
+        result += @rolls.fetch(frame_index, 0) + @rolls.fetch(frame_index + 1, 0)
+        frame_index += 2
+      end
     end
     result
   end
 
   private
-  def strike?(frame_index)
-    @rolls.fetch(frame_index, 0) == PINS
-  end
-  def strike_bonus(frame_index)
-    @rolls.fetch(frame_index + 1, 0) + @rolls.fetch(frame_index + 2, 0) + PINS
-  end
-
   def spare?(frame_index)
-    @rolls.fetch(frame_index,0) + @rolls.fetch(frame_index + 1,0) == PINS
+    @rolls.fetch(frame_index, 0) + @rolls.fetch(frame_index + 1, 0) == PINS
   end
   def spare_bonus(frame_index)
-    @rolls.fetch(frame_index + 2, 0) + PINS
+    PINS + @rolls.fetch(frame_index + 2, 0)
   end
-
 end
