--- spec/bowler_test.rb
@@ -1,10 +1,0 @@
-require 'minitest/autorun'
-require_relative '../lib/bowler'
-
-class BowlerTest < MiniTest::Test
-
-  def test_bowler_instance
-    bowler = Bowler.new
-    assert_instance_of Bowler, bowler
-  end
-end
